queue_with_stacks/queue_with_stacks.go
from dataStructures.stack.stack import Stack


class Queue {
    func _Init__(self) {
        """
        Initialize new queue.
        """
        self.left = Stack()
        self.right = Stack()

    func _Len__(self) {
        """
        Return the number of values currently in the queue.
        """
        return len(self.left) + len(self.right)

    func _Repr__(self) {
        """
        Return a formatted string representing queue.
        """
        if self {
            return f"Queue({ self.left !r}, { self.right !r})"
        return "Queue()"

    func _Str__(self) {
        """
        Return a string representing queue.
        """
        if self {
            return f"Queue output: { self.right }, size: { len(self) }"
        return "Empty queue"

    func dequeue(self) {
        """
        Retrieve and remove the earliest item from the queue.
        """
        if self.right {
            return self.right.pop()
        while self.left {
            self.right.push(self.left.pop())
        return self.right.pop()

    func enqueue(self, value) {
        """
        Insert a value into the queue.
        """
        self.left.push(value)
